site settings banner add, load, and update feture added

+++ modules/categories.js
@@ -3,17 +3,22 @@ import { ObjectId } from "mongodb";
 //get all cetegoris load
 export const getAllCategories = (categoriesCollection) => {
   return async (req, res) => {
-    const cursor = await categoriesCollection.find().toArray();
+    try{
+      const cursor = await categoriesCollection.find().toArray();
     res.send(cursor);
+    }
+    catch(err){
+      return res.status(500).send({ message: "Not Found", err });
+    }
   };
 };
 
 // add new categories
 export const postNewCategories = (categoriesCollection) => {
   return async (req, res) => {
-    const {categoryName} = req.body;
-    const addNewCategory = ['all', categoryName]; 
-   
+    const { categoryName } = req.body;
+    const addNewCategory = ["all", categoryName];
+
     const query = { categoryName: addNewCategory };
 
     const alreadeyAdded = await categoriesCollection.findOne(query);
@@ -23,7 +28,9 @@ export const postNewCategories = (categoriesCollection) => {
         message: "This category already added",
       });
     }
-    const newAddCategory = await categoriesCollection.insertOne({categoryName : addNewCategory});
+    const newAddCategory = await categoriesCollection.insertOne({
+      categoryName: addNewCategory,
+    });
     return res.send(newAddCategory);
   };
 };
@@ -35,32 +42,29 @@ export const putCategoryUpdate = (categoriesCollection, productsCollection) => {
   return async (req, res) => {
     try {
       const id = req.params.id;
-      const oldCategoryName = req.body.oldCategoryName;
-      const updateCategoryName = req.body.updateCategoryName;
-      console.log('update and old category name', oldCategoryName, updateCategoryName);
+      const { oldCategoryName, updateCategoryName } = req.body;
 
+     
       if (!id || !oldCategoryName || !updateCategoryName) {
         return res.status(400).send({ message: "Invalid input data" });
       }
 
-      const productUpdateResult = await productsCollection.updateMany(
-        { categoryName: oldCategoryName },
-        { $set: { "categoryName.$": updateCategoryName } }
+      await productsCollection.updateMany(
+        { productCategory: oldCategoryName },
+        { $set: { "productCategory.$": updateCategoryName } },
+        
       );
+      
+      
 
-      if (productUpdateResult.matchedCount === 0) {
-        return res
-          .status(404)
-          .send({ message: "No products found with the given category" });
-      } else if (productUpdateResult.modifiedCount === 0) {
-        return res.status(500).send({ message: "Failed to update products" });
-      }
-
-      const query = { _id: new ObjectId(id), categoryName : oldCategoryName };
+      const query = { _id: new ObjectId(id), categoryName: oldCategoryName };
       const updateDoc = {
         $set: { "categoryName.$": updateCategoryName },
       };
-      const options = { upsert: false };
+      const options = {
+        
+        upsert: false
+      };
 
       const categoryUpdateresult = await categoriesCollection.updateOne(
         query,
@@ -79,13 +83,36 @@ export const putCategoryUpdate = (categoriesCollection, productsCollection) => {
 };
 
 // category datele
-export const deleteCategoryOne = (categoriesCollection) => {
+export const deleteCategoryOne = (categoriesCollection, productsCollection) => {
   return async (req, res) => {
-    const id = req.params.id;
-    const query = { _id: new ObjectId(id) };
+    try{
+      const id = req.params.id;
+      const query = { _id: new ObjectId(id) };
+      
+      const category = await categoriesCollection.findOne(query)
+     
+     
+      if (!id || !category) {
+        return res.status(400).send({ message: "Invalid input data" });
+      }
 
-    // todo : when category delete, then similar category product set new categories.
-    const result = await categoriesCollection.deleteOne(query);
-    return res.send(result);
+       await productsCollection.updateMany(
+        { productCategory: category.categoryName },
+        { $set: {productCategory : ['all', 'empty'] } },
+        
+      );
+      
+      
+     
+
+     
+      const deleteResult = await categoriesCollection.deleteOne(query);
+      return res.send(deleteResult);
+
+    } 
+    catch(err){
+      return res.status(500).send({ message: "An error occurred", err });
+    }
+   
   };
 };
